double

+++ BowlingMVC/Models/Price.cs
@@ -3,8 +3,8 @@
     public class Price
     {
         public int PId { get; set; }
-        public int NormalPrice { get; set; }
-        public int SpeialPrice { get; set; }
+        public double NormalPrice { get; set; }
+        public double SpeialPrice { get; set; }
         public String Weekday { get; set; }
         public Price() { }
     }
